AR-02: Android 0.16.0 — Gradle 9.7.1, AGP 9.4.1, Kotlin 2.4.20
Die Bau-Sprache wechselt, die Bibliotheken stehen absichtlich still.
kotlin-android ist aus den Modulen ausgetragen (AGP 9 übersetzt Kotlin
selbst); gemeinsam/ hängt an kotlin.directories statt java.srcDir, das das
eingebaute Kotlin nicht mehr annimmt; testReleaseUnitTest bleibt über
android.onlyEnableUnitTestForTheTestedBuildType=false an — sonst entfiele
der Release-Fall von ServeradresseTest (Nr. 142) still.

Zwölf Veraltungen der Bauskripte umgestellt, nicht unterdrückt. Kotlin 2.4
fand vier Warnungen, darunter einen Fehler im Prüffall: KopplungRundlaufTest
gab den Rückstand als nachgestellte Lambda, die seit Nr. 114 an raeumen ging
(F-AR-11, ohne Wirkung, jetzt benannt).

Gemessen (mit 0.15.1 gebaut, damit der Stempel nicht mitzählt): 264/15 und
71 Prüffälle je Bauart wie vorher, 78 von 78 Bildern byteweise gleich,
Manifest und Ressourcen beider APKs gleich; verschieden nur DEX, Builtins
und Baseline-Profil (kotlin-stdlib 2.1.21 -> 2.4.20). Lint Handy 13 / Uhr 1,
0 Fehler, 0 Kotlin-Warnungen. Gradle-Prüfsumme zweifach, Wrapper-JAR aus
der geprüften Verteilung. Versionscodes am APK: 1600 / 1001600.

Nummer 0.16.0 für die ganze Runde (E-AR-04), Nebenstufe wegen targetSdk 37
(E-AR-08); sie trägt auch die Hausform-Zeile aus PK-04/5b (Nr. 284).

+++ android/handy/src/test/java/org/genem/nadoku/handy/HandyBildTest.kt
@@ -591,10 +591,10 @@ class HandyBildTest {
             breiteDp = breiteDp,
             knopfDp = if (band == null) 0.0 else (band.last - band.first + 1) / dichte.toDouble(),
             knopfUnterkanteDp =
-                if (band == null) 0 else Math.round((band.last + 1) / dichte).toInt(),
+                if (band == null) 0 else Math.round((band.last + 1) / dichte),
             amRand = knopfAmRand(punkte, breitePx, hoehePx),
             unterDerFaltkante = band != null && band.last == hoehePx - 1,
-            inhaltDp = Math.round((letzteZeile + 1) / dichte).toInt(),
+            inhaltDp = Math.round((letzteZeile + 1) / dichte),
             knoepfeGesamt = alleBaender.size,
             knoepfeSichtbar = alleBaender.count { it.last < hoehePx },
             farben = punkte.toHashSet().size,
